+++ src/components/courierInboxVue.ts
@@ -1,6 +1,6 @@
 import { h, type PropType, defineComponent } from 'vue-demi';
 
-import type { InboxPlacement, InboxTheme, Brand, Labels, InboxTrigger } from '../types/inbox';
+import type { InboxPlacement, InboxTheme, Brand, Labels, InboxTrigger, OnEvent } from '../types/inbox';
 
 type IsoFormatter = (isoDate: string) => string;
 
@@ -9,18 +9,28 @@ export default defineComponent({
     tenantId: String,
     brand: Object as PropType<Brand>,
     className: String,
-    openLinksInNewTab: Boolean,
-    renderActionsAsButtons: Boolean,
-    showUnreadMessageCount: Boolean,
+    defaultIcon: String,
     from: Number,
-    views: Array as PropType<{id: string, label: string, params: any}[]>,
+    isOpen: Boolean,
+    views: Array as PropType<{ id: string; label: string; params: any }[]>,
     formatDate: Function as PropType<IsoFormatter>,
+    appendTo: String,
     labels: Object as PropType<Labels>,
-    defaultIcon: { type: [Boolean, String] },
+    onEvent: Object as PropType<OnEvent>,
+    openLinksInNewTab: { type: Boolean, default: true },
     placement: Object as PropType<InboxPlacement>,
+    showUnreadMessageCount: Boolean,
     theme: Object as PropType<InboxTheme>,
     title: String,
     trigger: Object as PropType<InboxTrigger>
+    // The following cannot be easily converted from React.FunctionComponent to a vue equivalent
+    // renderContainer:
+    // renderBell:
+    // renderFooter:
+    // renderPin:
+    // renderIcon:
+    // renderMessage:
+    // renderNoMessages
   },
   setup(props) {
     return () => h('courier-inbox', { ...props });