+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 4.5.0
+- upgrade to courier v4.5.0
+- added support for `appendTo` to append inbox anywhere
+- added missing inbox support for `defaultIcon`, `isOpen`, `onEvent`, `openLinksInNewTab` and `showUnreadMessageCount`
+
 # 4.4.1
 - added support for bundler imports